fixed version compare
Comparing versions on a numeric base results in the unexpected truth that float(10.11) is smaller than (10.4).

This lead to the wrong queries being used on newer mysql/maria db installations (>=10.10) as the comparison flag was being set to false.

To achieve a version comparison which respects version strings and their ordering the python package "packaging" is used. This way the version comparison works as expected and now the right queries are used.